product controller update, orderdto added

+++ FoodOrderApp/Controllers/Api/OrderController.cs
@@ -1,6 +1,9 @@
 using FoodOrderApp.Data.Abstract;
+using FoodOrderApp.DTOs;
 using FoodOrderApp.Entity;
 using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace FoodOrderApp.Controllers
@@ -16,6 +19,26 @@ namespace FoodOrderApp.Controllers
             _orderRepository = orderRepository;
         }
 
+        private OrderDto MapToDto(Order order)
+        {
+            return new OrderDto
+            {
+                OrderId = order.OrderId,
+                OrderDate = order.OrderDate,
+                OrderMethod = order.OrderMethod,
+                CardId = order.CardId,
+                AddressId = order.AddressId
+            };
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> GetOrders()
+        {
+            var orders = await _orderRepository.GetAllAsync();
+            var orderDtos = orders.Select(MapToDto);
+            return Ok(orderDtos);
+        }
+
         [HttpGet("{id}")]
         public async Task<IActionResult> GetOrder(int id)
         {
@@ -24,23 +47,43 @@ namespace FoodOrderApp.Controllers
             {
                 return NotFound();
             }
-            return Ok(order);
+            return Ok(MapToDto(order));
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateOrder([FromBody] Order order)
+        public async Task<IActionResult> CreateOrder([FromBody] OrderDto orderDto)
         {
+            var order = new Order
+            {
+                OrderDate = orderDto.OrderDate,
+                OrderMethod = orderDto.OrderMethod,
+                CardId = orderDto.CardId,
+                AddressId = orderDto.AddressId
+            };
+
             await _orderRepository.AddAsync(order);
-            return CreatedAtAction(nameof(GetOrder), new { id = order.OrderId }, order);
+            var createdOrderDto = MapToDto(order);
+            return CreatedAtAction(nameof(GetOrder), new { id = order.OrderId }, createdOrderDto);
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateOrder(int id, [FromBody] Order order)
+        public async Task<IActionResult> UpdateOrder(int id, [FromBody] OrderDto orderDto)
         {
-            if (id != order.OrderId)
+            if (id != orderDto.OrderId)
             {
                 return BadRequest();
             }
+
+            var order = await _orderRepository.GetByIdAsync(id);
+            if (order == null)
+            {
+                return NotFound();
+            }
+
+            order.OrderDate = orderDto.OrderDate;
+            order.OrderMethod = orderDto.OrderMethod;
+            order.CardId = orderDto.CardId;
+            order.AddressId = orderDto.AddressId;
 
             await _orderRepository.UpdateAsync(order);
             return NoContent();
@@ -49,15 +92,14 @@ namespace FoodOrderApp.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteOrder(int id)
         {
+            var order = await _orderRepository.GetByIdAsync(id);
+            if (order == null)
+            {
+                return NotFound();
+            }
+
             await _orderRepository.DeleteAsync(id);
             return NoContent();
-        }
-
-        [HttpGet]
-        [Route("admin/orders")]
-        public async Task<IActionResult> GetOrders()
-        {
-            return Ok(await _orderRepository.GetAllAsync());
         }
     }
 }
